Adiciona classe genérica para ser herdada pelas subclasses
